Move tinyformat.h to utility subdir
