+++ configure.ac
@@ -313,12 +313,12 @@ AC_CONFIG_LINKS([include/souffle/utility/MiscUtil.h:src/utility/MiscUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/StreamUtil.h:src/utility/StreamUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/StringUtil.h:src/utility/StringUtil.h])
 AC_CONFIG_LINKS([include/souffle/utility/ParallelUtil.h:src/utility/ParallelUtil.h])
+AC_CONFIG_LINKS([include/souffle/utility/tinyformat.h:src/utility/tinyformat.h])
 AC_CONFIG_LINKS([include/souffle/WriteStream.h:src/WriteStream.h])
 AC_CONFIG_LINKS([include/souffle/WriteStreamCSV.h:src/WriteStreamCSV.h])
 AC_CONFIG_LINKS([include/souffle/WriteStreamSQLite.h:src/WriteStreamSQLite.h])
 AC_CONFIG_LINKS([include/souffle/gzfstream.h:src/gzfstream.h])
 AC_CONFIG_LINKS([include/souffle/json11.h:src/json11.h])
-AC_CONFIG_LINKS([include/souffle/tinyformat.h:src/tinyformat.h])
 AC_CONFIG_LINKS([include/souffle/profile/Cell.h:src/profile/Cell.h])
 AC_CONFIG_LINKS([include/souffle/profile/CellInterface.h:src/profile/CellInterface.h])
 AC_CONFIG_LINKS([include/souffle/profile/Cli.h:src/profile/Cli.h])
