+++ src/utility/MiscUtil.h
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include "../tinyformat.h"
+#include "tinyformat.h"
 #include <cassert>
 #include <chrono>
 #include <cstdlib>
